Split chapel-py method generation into several files
