+++ tools/chapel-py/src/param-methods.cpp
@@ -1,0 +1,24 @@
+/*
+ * Copyright 2023-2026 Hewlett Packard Enterprise Development LP
+ * Other additional copyright holders may be indicated within.
+ *
+ * The entirety of this work is licensed under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ *
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/* this sets up the necessaary macros to code-generate methods from method tables. */
+#include "core-types-gen-help.h"
+
+/* This invokes the macros, performing the code-generation. */
+#include "method-tables/param-methods.h"
